split dm_spectral into dmpipe/dm_spectral_utils.py
